Fix moderation cleanup callback export
