+++ demo/video-chat/frontend-vue/tests/contract/refactor-commit-boundaries-contract.mjs
@@ -59,6 +59,7 @@ try {
     'lifecycle.js',
     'mediaSecurityRuntime.js',
     'mediaStack.js',
+    'moderationSync.js',
     'nativeStack.js',
     'orchestration.js',
     'participantUi.js',
@@ -73,6 +74,13 @@ try {
   for (const fileName of requiredHelperFiles) {
     assert.ok(fs.existsSync(path.resolve(workspaceHelpersRoot, fileName)), `workspace helper file missing: ${fileName}`);
   }
+
+  assert.ok(!workspaceView.includes('moderationSyncQueue'), 'CallWorkspaceView.vue must not own moderation sync queues');
+  assert.ok(!workspaceView.includes('let moderationSyncTimer'), 'CallWorkspaceView.vue must not own moderation sync timers');
+
+  const participantUi = read('src/domain/realtime/workspace/callWorkspace/participantUi.js');
+  assert.ok(participantUi.includes("import { createCallWorkspaceModerationSync } from './moderationSync';"), 'participant UI must delegate moderation sync queue handling');
+  assert.ok(participantUi.includes('consumeQueuedModerationSyncEntries,'), 'participant UI must expose consumeQueuedModerationSyncEntries for lifecycle cleanup');
 
   const helperFiles = fs.readdirSync(workspaceHelpersRoot).filter((name) => name.endsWith('.js') && !name.endsWith('.extracted.js'));
   assert.ok(helperFiles.length >= 12, `workspace helper surface must remain modular (got ${helperFiles.length} helper modules)`);
